Kludges for sockets to seem to work
File descriptors on windows can not be simultaneously used from multiple
threads - read and write grab a lock on file descriptor.

So, instead of calling read and write, handles are obtained from FD and
ReadFile and WriteFile functions are used instead.

It seems that on windows if one thread calls ReadFile and blocks on it,
the other thread can't call WriteFile on the same handle - it would
block until ReadFile returns. This happens at least for sockets;
seems to be some kind of internal lock on a socket.

So, win32_unix_read busy-loops checking for input on a socket (by ioctl
FIONREAD) and only then calls ReadFile.